Allow plugins for the LTI XBlock to pass extra parameters to the provider
Overview
---------
This pull request makes this XBlock extensible for
adding parameters to the LTI provider in addition
to the existing ones.

Why
---
A customer needs `team` and `cohort` informaiton
to be passed through the LTI channel to simplify
integration with their LTI provider. At the same
time we'd rather not fork the XBlock, so an
Open-Closed model is needed.

More Detailed Documentation
---------------------------
I've added a couple of sections to `README.rst`
to document specifying design and
configuration steps of the new LTI parameter
processors. Please check it out for more
information about the changes that this pull
request introduces.

Example Usage
-------------
I've written a package that we'll need to use for
our customer to pass some personal user
information in addition to the cohort and team
info: https://github.com/appsembler/tahoe-lti/pull/1

Omar Al-Ithawi @ Appsembler