fix: Manually backfill lti_config dicts
The difficulty here is that the empty string in an invalid JSON
dictionary (as opposed to None/NULL or '{}' which is just "blank");
attempts to instantiate models would fail, given the invalid data. This
meant we couldn't use the Django ORM to handle the migration entirely;
we need to craft some raw SQL to work-around these checks and
limitations.

Related-To: 42a9e342ef57a9ffa5016f4f57da633a811d2c0f
Related-To: https://github.com/edx/xblock-lti-consumer/pull/144