refactor: make access token endpoint backwards compatible
This commit makes the access token endpoint backwards compatible with
the XBlock handler. This is achieved by saving the LTI 1.3 configuration
to the Django model in the XBlock handler and calling the view which
loads this config and generates the access token.

The LTIConfiguration model's LTI 1.3 consumer function is updated to
support CONFIG_ON_DB to support passing the configuration from the
XBlock to the Django View.