refactor: make LTI 1.3 launch callback backwards compatible
The logic of the LTI 1.3 Launch callback has been moved from XBlock to
the plugin view. This commit updates the XBlock handler to be backwards
compatible with the previous functionality and updates the relevant
tests. The XBlock handler imports the Django view in runtime and calls
it as function using the request object and returns the result.