feat: adds access token view for backward compatibility
Changing the URL of the access token endpoint to use the config id
instead of the XBlock location helps in decoupling LTI from XBlocks. But
any existing integration would be broken because the access token URL
generated in previous versions use the XBlock location.

This commit re-adds the access token endpoint that calls the XBlock
handler to provide backward compatibility with the URLs generated in the
previous versions of LTI Consumer.