fix: change user from None to AnonymousUser in authenticated requests
Modified the Lti1p3ApiAuthentication authentication backend to return
AnonymousUser instead of None. This allows DarkLangMiddleware to work
properly without crashing when middleware tries to get a user info
from the request.

This investigation and patch was originally done by @OlhaShyliaieva in:
  https://github.com/openedx/xblock-lti-consumer/pull/228

This is a Nutmeg backport of the 4.2.2 release. See:
  https://github.com/openedx/xblock-lti-consumer/pull/261

I'm just rebasing it onto the latest version to land this change.