feat: replace use of CourseWaffleFlag with WaffleFlag for database_config_waffle_flag
Support for CONFIG_ON_DATABASE in LTI 1.3 was added recently. These changes were put behind a CourseWaffleFlag. However, the CourseWaffleFlag.is_enabled(<course_key>) method must be passed a course_key as an argument. This method is used within the codebase to determine whether CONFIG_ON_DATABASE support for LTI 1.3 is enabled.

Outside of the context of the edx-platform, CourseWaffleFlag does not exist. Other consumers of this library that are not the edx-platform will encounter errors when working with the library, because the location field of LtiConfiguration may be None. In this case, there is no way to get the course_key from the XBlock.

This commit changes the CourseWaffleFlag to a WaffleFlag so that a course_key is not necessary.

The alternative is to only use the flag when LtiConfiguration.location is not None, but that is not satisfactory. This change sacrifies increased customizability of the flag with better coverage for consumers of this library.