feat: add full name as a LTI PII field
This commits adds the ability to share a learner's full name in both LTI 1.1 and LTI 1.3 launches. In LTI 1.1, the full name is sent as the "lis_person_name_full" LTI parameter. In LTI 1.3, the full name is sent as the "name" ID token claim.

The full name corresponds to the learner's name as recorded by the UserProfile model in the LMS.

Enabling the transmission of full name in either version of LTI follows the existing pattern for PII sharing. In order for full name to be shared via an LTI launch, the CourseAllowPIISharingInLTIFlag configuration model must be enabled for the course, and the "Request user's full name" XBlock field must be enabled in the XBlock describing the particular LTI integration.

Please note that this commit does not add the ability to share a learner's full name via the LTI 1.3 Names and Role Provisioning Services context membership endpoint because that behavior already exists.